Add gradient border button as a storybook component and normalize component API
