+++ src/app/_components/_sharedcomponents/_styledcomponents/GradientBorderButton.tsx
@@ -1,36 +1,33 @@
-import { Button } from '@mui/material';
+import { Button, SxProps, ButtonProps as MuiButtonProps } from '@mui/material';
+import { Theme } from '@mui/material/styles';
 
-interface IGradientBorderButtonProps {
+export type IGradientBorderButtonProps = Omit<MuiButtonProps, 'sx'> & {
     fillColor?: string;
-    onClickHandler: () => void;
-    children: React.ReactNode;
-}
-
-const GradientBorderButton: React.FC<IGradientBorderButtonProps> = ({ fillColor = '#1E2D32', onClickHandler, children }) => {
-    return (
-        <Button
-            sx={{
-                borderRadius: '15px',
-                backgroundColor: '#1E2D32',
-                padding: '1rem 1.5rem',
-
-                border: '2px solid transparent',
-                background:
-                `linear-gradient(${fillColor}, ${fillColor}) padding-box, linear-gradient(to top, #038FC3, #595A5B) border-box`,
-                '&:hover': {
-                    filter: 'brightness(1.2)',
-                },
-                '&:disabled': {
-                    color: '#666666',
-                    cursor: 'default',
-                },
-                cursor: 'pointer',
-            }}
-            onClick={onClickHandler}
-        >
-            {children}
-        </Button>
-    );
+    sx?: SxProps<Theme>;
 };
+
+const GradientBorderButton: React.FC<IGradientBorderButtonProps> = ({ fillColor = '#1E2D32', sx, ...otherProps }) => (
+    <Button
+        sx={{
+            borderRadius: '15px',
+            backgroundColor: '#1E2D32',
+            padding: '1rem 1.5rem',
+
+            border: '2px solid transparent',
+            background:
+                `linear-gradient(${fillColor}, ${fillColor}) padding-box, linear-gradient(to top, #038FC3, #595A5B) border-box`,
+            '&:hover': {
+                filter: 'brightness(1.2)',
+            },
+            '&:disabled': {
+                color: '#666666',
+                cursor: 'default',
+            },
+            cursor: 'pointer',
+            ...sx
+        }}
+        {...otherProps}
+    />
+);
 
 export default GradientBorderButton;
